move chrome profile into promptlogger
