+++ runner/daemon/src/mill/daemon/MillMain0.scala
@@ -8,7 +8,7 @@ import mill.bsp.BSP
 import mill.client.lock.Lock
 import mill.constants.{DaemonFiles, OutFiles}
 import mill.api.BuildCtx
-import mill.internal.{Colors, MultiStream, PrefixLogger, PromptLogger, SimpleLogger}
+import mill.internal.{Colors, JsonArrayLogger, MultiStream, PrefixLogger, PromptLogger, SimpleLogger}
 import mill.server.Server
 import mill.util.BuildInfo
 import mill.api
@@ -292,7 +292,8 @@ object MillMain0 {
                               .orElse(Option.when(config.disableTicker.value)(false)),
                             daemonDir,
                             colored = colored,
-                            colors = colors
+                            colors = colors,
+                            out = out
                           )) { logger =>
                             proceed(logger)
                           }
@@ -521,7 +522,8 @@ object MillMain0 {
       enableTicker: Option[Boolean],
       daemonDir: os.Path,
       colored: Boolean,
-      colors: Colors
+      colors: Colors,
+      out: os.Path
   ): Logger & AutoCloseable = {
     new PromptLogger(
       colored = colored,
@@ -533,7 +535,8 @@ object MillMain0 {
       debugEnabled = config.debugLog.value,
       titleText = config.leftoverArgs.value.mkString(" "),
       terminfoPath = daemonDir / DaemonFiles.terminfo,
-      currentTimeMillis = () => System.currentTimeMillis()
+      currentTimeMillis = () => System.currentTimeMillis(),
+      chromeProfileLogger = new JsonArrayLogger.ChromeProfile(out / OutFiles.millChromeProfile)
     )
   }
